pls-pls-pls/requirements.txt
Flask==0.10.1
tox==1.7.2
twilio==3.6.8
python-forecastio==1.3.1
geopy==1.8.1
googlemaps==2.1